+++ include/hsmcpp/HsmEventDispatcherGLib.hpp
@@ -7,15 +7,19 @@
 #include "HsmEventDispatcherBase.hpp"
 #include <glib.h>
 #include <condition_variable>
+#include <map>
 
 namespace hsmcpp
 {
 
 class HsmEventDispatcherGLib: public HsmEventDispatcherBase
 {
+private:
+    using TimerData_t = std::pair<HsmEventDispatcherGLib*, TimerID_t>;
+
 public:
-    HsmEventDispatcherGLib();
-    explicit HsmEventDispatcherGLib(GMainContext* context);
+    explicit HsmEventDispatcherGLib(const size_t eventsCacheSize = DISPATCHER_DEFAULT_EVENTS_CACHESIZE);
+    HsmEventDispatcherGLib(GMainContext* context, const size_t eventsCacheSize = DISPATCHER_DEFAULT_EVENTS_CACHESIZE);
     virtual ~HsmEventDispatcherGLib();
 
     virtual void emitEvent(const HandlerID_t handlerID) override;
@@ -23,6 +27,14 @@ public:
     virtual bool start() override;
 
 private:
+    void unregisterAllTimerHandlers();
+
+    void startTimerImpl(const TimerID_t timerID, const unsigned int intervalMs, const bool isSingleShot) override;
+    void stopTimerImpl(const TimerID_t timerID) override;
+
+    static gboolean onTimerEvent(const TimerData_t* timerData);
+    static void onFreeTimerData(void* timerData);
+
     void notifyDispatcherAboutEvent() override;
     static gboolean onPipeDataAvailable(GIOChannel* gio, GIOCondition condition, gpointer data);
 
@@ -36,6 +48,7 @@ private:
     bool mDispatchingIterationRunning = false;
     std::mutex mDispatchingSync;
     std::condition_variable mDispatchingDoneEvent;
+    std::map<TimerID_t, GSource*> mNativeTimerHandlers;// <timerID, nativeTimerID>
 };
 
 } // namespace hsmcpp