+++ src/os/arduino/InterruptsFreeSection.cpp
@@ -1,15 +1,16 @@
 // Copyright (C) 2023 Igor Krechetov
 // Distributed under MIT license. See file LICENSE for details
-#include "hsmcpp/os/common/CriticalSection.hpp"
+#include "hsmcpp/os/common/InterruptsFreeSection.hpp"
 
 #include <Arduino.h>
 
 namespace hsmcpp {
-CriticalSection::CriticalSection() {
+
+InterruptsFreeSection::InterruptsFreeSection() {
     noInterrupts();  // disable interrupts
 }
 
-CriticalSection::~CriticalSection() {
+InterruptsFreeSection::~InterruptsFreeSection() {
     interrupts();  // enable interrupts
 }