+++ src/os/posix/InterruptsFreeSection.cpp
@@ -1,9 +1,10 @@
 // Copyright (C) 2023 Igor Krechetov
 // Distributed under MIT license. See file LICENSE for details
-#include "hsmcpp/os/posix/CriticalSection.hpp"
+#include "hsmcpp/os/posix/InterruptsFreeSection.hpp"
 
 namespace hsmcpp {
-CriticalSection::CriticalSection() {
+    
+InterruptsFreeSection::InterruptsFreeSection() {
     sigset_t blockMask;
 
     // add all signals to the set
@@ -12,7 +13,7 @@ CriticalSection::CriticalSection() {
     pthread_sigmask(SIG_BLOCK, &blockMask, &mOriginalSigMask);
 }
 
-CriticalSection::~CriticalSection() {
+InterruptsFreeSection::~InterruptsFreeSection() {
     // restore signals mask
     pthread_sigmask(SIG_SETMASK, &mOriginalSigMask, nullptr);
 }